folder_comparison_tool/requirements.txt
flask
python-dotenv
pandas
openpyxl
requests
groq